requirements.txt
langchain-core>=0.3
langchain-community>=0.3
langchain-openai>=0.3
langchain-chroma>=0.1.2
langchain-text-splitters>=0.3
langgraph>=0.0.37
langsmith>=0.1.100

chromadb>=0.5
rank_bm25>=0.2.2
sentence-transformers>=3.0.0

fastapi>=0.110
uvicorn>=0.27
httpx>=0.27

sqlalchemy>=2.0
pydantic>=2.0
python-dotenv>=1.0

streamlit>=1.32
openai>=1.0

# Optional MLOps (kept but not required to run)
mlflow>=2.9.2
wandb>=0.16

pytest>=8.0
ruff>=0.4.0
